requirements.txt
httpx==0.27.2
openai>=1.40.0
python-dotenv==1.0.1
jinja2==3.1.4
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
flask==3.0.0
requests==2.31.0
langchain==0.1.0
langchain-openai==0.0.5
gunicorn==21.2.0
